Move changes API classes to separate package, move other content to core sub-package and make exchange_calendars_Extensions a namespace package.
